requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
oauthlib==3.3.1
packaging==25.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
